+++ dashboard/admin.py
@@ -1,7 +1,1 @@
 from django.contrib import admin
-from dashboard.models import Dashboard
-
-
-@admin.register(Dashboard)
-class DashboardAdmin(admin.ModelAdmin):
-    pass
