+++ patient/apps.py
@@ -1,5 +1,5 @@
 from django.apps import AppConfig
 
 
-class SubmissionConfig(AppConfig):
-    name = 'submission'
+class PatientConfig(AppConfig):
+    name = 'patient'
