+++ patient/models.py
@@ -1,5 +1,4 @@
 import hashlib
-import uuid
 
 from django.core.validators import MaxValueValidator, MinValueValidator
 from django.db import models
